Add KiiAppTest
